requirements.txt
aiogram==3.6.0
httpx[http2]==0.27.0
pydantic==2.6.3
pydantic-settings==2.2.1
python-dotenv==1.0.1
structlog==24.1.0
tenacity==8.2.3
pytest==8.3.3
pytest-asyncio==0.23.5
respx==0.20.2


